Refactor plan selection
