+++ client/landing/stepper/declarative-flow/internals/steps-repository/sensei-plan/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable wpcalypso/jsx-classname-namespace */
 import { useLocale } from '@automattic/i18n-utils';
 import { useSelect } from '@wordpress/data';
-import { useCallback, useState, useEffect } from '@wordpress/element';
+import { useCallback, useState } from '@wordpress/element';
 import { sprintf } from '@wordpress/i18n';
 import { useI18n } from '@wordpress/react-i18n';
 import classnames from 'classnames';
@@ -11,28 +11,20 @@ import PlanItem from 'calypso/../packages/plans-grid/src/plans-table/plan-item';
 import FormattedHeader from 'calypso/components/formatted-header';
 import { ONBOARD_STORE } from 'calypso/landing/stepper/stores';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
-import { domainRegistration } from 'calypso/lib/cart-values/cart-items';
-import { cartManagerClient } from 'calypso/my-sites/checkout/cart-manager-client';
 import PlanPrice from 'calypso/my-sites/plan-price';
 import { SenseiStepContainer } from '../components/sensei-step-container';
 import { SenseiStepError } from '../components/sensei-step-error';
 import { SenseiStepProgress } from '../components/sensei-step-progress';
 import { PlansIntervalToggle } from './components';
 import { useFeatures, Status } from './constants';
-import { useCreateSenseiSite } from './create-sensei-site';
 import { useBusinessPlanPricing, useSenseiProPricing } from './sensei-plan-products';
+import { usePlanSelection } from './use-plan-selection';
 import type { Step } from '../../types';
-import type { NewSiteBlogDetails, OnboardSelect } from '@automattic/data-stores';
-import type { MinimalRequestCartProduct } from '@automattic/shopping-cart';
+import type { OnboardSelect } from '@automattic/data-stores';
 import type { PlanBillingPeriod } from 'calypso/../packages/data-stores';
 
 import 'calypso/../packages/plans-grid/src/plans-table/style.scss';
 import './styles.scss';
-
-const isComingFromSenseiLMSSite = () => {
-	const urlParams = new URLSearchParams( window.location.search );
-	return urlParams.get( 'ref' ) === 'senseilms';
-};
 
 const SenseiPlan: Step = ( { flow, navigation: { submit } } ) => {
 	const [ billingPeriod, setBillingPeriod ] = useState< PlanBillingPeriod >( 'ANNUALLY' );
@@ -55,77 +47,15 @@ const SenseiPlan: Step = ( { flow, navigation: { submit } } ) => {
 		submit?.( undefined, 'senseiDomain' );
 	}, [ submit ] );
 
-	const { createAndConfigureSite, progress } = useCreateSenseiSite();
-
-	const createCheckoutCart = useCallback(
-		async ( site?: NewSiteBlogDetails ) => {
-			const cartKey = await cartManagerClient.getCartKeyForSiteSlug( site?.site_slug ?? '' );
-
-			const productsToAdd: MinimalRequestCartProduct[] = [
-				{
-					product_slug: businessPlan.productSlug,
-					extra: {
-						signup_flow: flow,
-					},
-				},
-				{
-					product_slug: senseiProPlan.productSlug,
-					extra: {
-						signup_flow: flow,
-					},
-				},
-			];
-
-			if ( domain && domain.product_slug ) {
-				const registration = domainRegistration( {
-					domain: domain.domain_name,
-					productSlug: domain.product_slug,
-					extra: { privacy_available: domain.supports_privacy },
-				} );
-
-				productsToAdd.push( registration );
-			}
-
-			await cartManagerClient.forCartKey( cartKey ).actions.addProductsToCart( productsToAdd );
-			const redirectTo = encodeURIComponent(
-				`/setup/sensei/senseiPurpose?siteSlug=${ site?.site_slug }&siteId=${ site?.blogid }`
-			);
-
-			return `/checkout/${ site?.site_slug }?signup=1&redirect_to=${ redirectTo }`;
-		},
-		[ businessPlan.productSlug, domain, flow, senseiProPlan.productSlug ]
-	);
-
-	const onPlanSelect = useCallback( async () => {
-		try {
-			setStatus( Status.Bundling );
-
-			// Wait for a bit to get an animation in the beginning of site creation.
-			await new Promise( ( res ) => setTimeout( res, 100 ) );
-
-			const { site } = await createAndConfigureSite();
-			const checkoutUrl = await createCheckoutCart( site );
-
-			return window.location.assign( checkoutUrl );
-		} catch ( err ) {
-			setStatus( Status.Error );
-		}
-	}, [ createAndConfigureSite, createCheckoutCart ] );
-
-	// Auto-select YEARLY plan when coming from senseilms.com site.
-	useEffect( () => {
-		if ( ! isComingFromSenseiLMSSite() ) {
-			return;
-		}
-
-		if ( status === Status.Initial ) {
-			setStatus( Status.Bundling );
-		}
-
-		if ( ! isLoading ) {
-			onPlanSelect();
-		}
-	}, [ isLoading, status, onPlanSelect ] );
+	const { onPlanSelect, progress } = usePlanSelection( {
+		flow,
+		status,
+		setStatus,
+		domain,
+		businessPlanSlug: businessPlan.productSlug,
+		senseiProPlanSlug: senseiProPlan.productSlug,
+		isLoading,
+	} );
 
 	const currencyCode = senseiProPlan.currencyCode;
 	const price = businessPlan.price + senseiProPlan.price;
